requirements.txt
kafka-python
praw
pandas
streamlit
nltk
matplotlib
wordcloud
vaderSentiment
